scripts: tooling: tt-console: support rootless rescan via IOCTL
Support rootless rescan of the PCIe bus using an IOCTL- if this rescan
fails, we will fall back to a standard rescan using the PCIe sysfs
(which does require root access)
