+++ scripts/tooling/Makefile
@@ -12,7 +12,7 @@ all: $(notdir $(TOOLS))
 
 CFLAGS += -O0 -g -Wall -Wextra -Werror -std=gnu11 -I ../../include
 
-tt-console: console.c vuart.c
+tt-console: console.c vuart.c rescan.c
 	$(CC) $(CFLAGS) -o $(OUTDIR)/$@ $^
 
 tt-tracing: tracing.c vuart.c
